requirements.txt
#CUDA11.8に対応したtorchのインストールのurlを指定する。cudaのバージョンによって変更する必要がある。cuda使わない場合はfind-linksを削除する
--find-links https://download.pytorch.org/whl/cu118/torch_stable.html

numpy==1.26.4 #numpy2.0.0以上だと現在のところwandbが動かない
torch==2.3.1
torchaudio
torchvision
torchinfo
torchmetrics==0.11.4
einops
mne==1.6.0
hydra-core==1.3.2
wandb==0.15.0
termcolor
matplotlib
#下はCLIPを使う場合に必要なライブラリ
tqdm
regex
ftfy
git+https://github.com/openai/CLIP.git

timm